Rename file from services to service
